+++ src/contexts/Transactions/domain/payment-split.valueobject.ts
@@ -1,3 +1,4 @@
+import { NumberValueObject } from "@juandardilag/value-objects";
 import { AccountID } from "contexts/Accounts/domain";
 import { TransactionAmount } from "./transaction-amount.valueobject";
 
@@ -46,9 +47,18 @@ export class PaymentSplit {
 		);
 	}
 
-	static totalAmount(splits: PaymentSplit[]): TransactionAmount {
+	static totalAmount(
+		splits: PaymentSplit[],
+		exchangeRate?: NumberValueObject
+	): TransactionAmount {
 		return new TransactionAmount(
-			splits.reduce((sum, split) => sum + split.amount.value, 0)
+			splits.reduce(
+				(sum, split) =>
+					sum +
+					split.amount.times(exchangeRate ?? new NumberValueObject(1))
+						.value,
+				0
+			)
 		);
 	}
 }